+++ tests/lit/driver/lit.cfg.py
@@ -6,6 +6,8 @@ config.test_format = lit.formats.ShTest(execute_external=True)
 # Recognize .silly files as tests
 config.suffixes = ['.silly']
 
+config.excludes = ['mod1.silly', 'mod2.silly']
+
 # Use the generated site config
 config.test_source_root = os.path.dirname(__file__)
 config.test_exec_root = os.path.join(config.test_source_root, "..", "..", "..", "build", "tests", "lit", "driver")